requirements.txt
streamlit
supabase
psycopg2-binary
python-dotenv
pandas